Fix dependency between re_time_panel & re_viewport
