+++ crates/re_time_panel/Cargo.toml
@@ -19,6 +19,7 @@ all-features = true
 re_arrow_store.workspace = true
 re_data_store.workspace = true
 re_data_ui.workspace = true
+re_format.workspace = true
 re_log_types.workspace = true
 re_ui.workspace = true
 re_viewer_context.workspace = true
@@ -26,7 +27,6 @@ re_viewer_context.workspace = true
 egui.workspace = true
 itertools.workspace = true
 serde = "1.0"
-time.workspace = true
 vec1 = "1.8"
 
 [target.'cfg(not(target_arch = "wasm32"))'.dependencies]
